Handle exceptions

+++ src/controlER/encryption.py
@@ -29,22 +29,28 @@ class Encryption:
         """
         # Joining the path and filename together.
         file_path = os.path.join(path, filename)
+        if not os.path.exists(file_path):
+            print("File does not exists to be encrypted")
+            return
         #calculate name
         name = filename.split('.')[0]
-        # generating iv and send it
-        iv = getRandomNBitInteger(counter_size)
-        self.communication.add_value(iv, f'iv_{name}')
-        # Creating a counter object with the given counter size and initial value.
-        counter = Counter.new(counter_size, initial_value=iv)
-        # Creating a new AES object with the given key, mode and counter.
-        aes = AES.new(self.key, mode=AES.MODE_CTR, counter=counter)
-        # Opening the file, reading the data, encrypting it and then closing the file.
-        with open(file_path, 'rb') as file:
-            csv_data = file.read()
-            encrypted_data = aes.encrypt(csv_data)
-            file.close()
-        # Sending the encrypted data
-        self.communication.write_bytes(encrypted_data, name)
+        try:
+            # generating iv and send it
+            iv = getRandomNBitInteger(counter_size)
+            self.communication.add_value(iv, f'iv_{name}')
+            # Creating a counter object with the given counter size and initial value.
+            counter = Counter.new(counter_size, initial_value=iv)
+            # Creating a new AES object with the given key, mode and counter.
+            aes = AES.new(self.key, mode=AES.MODE_CTR, counter=counter)
+            # Opening the file, reading the data, encrypting it and then closing the file.
+            with open(file_path, 'rb') as file:
+                csv_data = file.read()
+                encrypted_data = aes.encrypt(csv_data)
+                file.close()
+            # Sending the encrypted data
+            self.communication.write_bytes(encrypted_data, name)
+        except:
+            print("Failed to encrypt file")
 
     def encrypt_authenticated(self, filename, path):
         """
@@ -53,20 +59,26 @@ class Encryption:
         :param filename: the name of the file to be encrypted
         :param path: the path to the file
         """
-        #calculate name
-        name = filename.split('.')[0]
         # Joining the path and filename together.
         file_path = os.path.join(path, filename)
-        # The nonce is a random number and sending it
-        nonce = get_random_bytes(12)
-        self.communication.write_bytes(nonce, f'nonce_{name}')
-        # Creating a new AESGCM object with the given key.
-        aesgcm = AESGCM(self.key)
-        # Opening the file, reading the data, encrypting it and then closing the file.
-        with open(file_path, 'rb') as file:
-            csv_data = file.read()
-            encrypted_data = aesgcm.encrypt(nonce, csv_data, None)
-            file.close()
+        if not os.path.exists(file_path):
+            print("File does not exists to be encrypted")
+            return
+        #calculate name
+        name = filename.split('.')[0]
+        try:
+            # The nonce is a random number and sending it
+            nonce = get_random_bytes(12)
+            self.communication.write_bytes(nonce, f'nonce_{name}')
+            # Creating a new AESGCM object with the given key.
+            aesgcm = AESGCM(self.key)
+            # Opening the file, reading the data, encrypting it and then closing the file.
+            with open(file_path, 'rb') as file:
+                csv_data = file.read()
+                encrypted_data = aesgcm.encrypt(nonce, csv_data, None)
+                file.close()
 
-        # Sending the encrypted data
-        self.communication.write_bytes(encrypted_data, name)
+            # Sending the encrypted data
+            self.communication.write_bytes(encrypted_data, name)
+        except:
+            print("Failed to encrypt file")
